+++ src/constant/argument.js
@@ -1,4 +1,5 @@
 import {
+  EMAIL_LEAKS,
   SCAN,
   HASH,
   DNS,
@@ -9,20 +10,20 @@ import {
 const mode = ['detailed', 'simple', 'blacklist', 'geoip']
 
 const argument = {
-  // [EMAIL_LEAKS]: [
-  //   {
-  //     arg: 'email',
-  //     required: true,
-  //     check: {
-  //       func: 'isEmail',
-  //       args: [{
-  //         require_display_name: false,
-  //         allow_display_name: false,
-  //         allow_utf8_local_part: true
-  //       }]
-  //     }
-  //   }
-  // ],
+  [EMAIL_LEAKS]: [
+    {
+      arg: 'email',
+      required: true,
+      check: {
+        func: 'isEmail',
+        args: [{
+          require_display_name: false,
+          allow_display_name: false,
+          allow_utf8_local_part: true
+        }]
+      }
+    }
+  ],
   [SCAN]: [
     {
       arg: 'ip',