+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+2.0.0
+* Added bower support
+* UMD Module fix
+* Queue added to TmLabs class.
+* TmLabs class init options added limit. Queue concurrency option value
+* Added batch commands for TmLabs class
+* Email Leaks endpoint added
+
 1.0.0
 * Initial release